Seperate idlemodule into its own file
